Feat: implement track-resonance and track-cascade pair task

+++ PWGCF/Femto/Core/dataTypes.h
@@ -25,7 +25,7 @@ namespace femtodatatypes
 // Note: Length of the bitmask is the limit of how many selections can be configured
 
 // datatypes for collsions
-using CollisionTagType = __uint128_t;
+using CollisionTagType = uint64_t;
 using CollisionMaskType = uint16_t;
 
 // datatypes for tracks
